+++ lib/generators/spree_emerchantpay_genesis/install/install_generator.rb
@@ -1,3 +1,5 @@
+require 'rails/generators' # Fix Rails::Generators::Base missing class error in earlier Spree versions
+
 module SpreeEmerchantpayGenesis
   module Generators
     # Emerchantpay plugin install script