+++ spec/helpers/spree_emerchantpay_genesis/transaction_helper_spec.rb
@@ -2,7 +2,7 @@ RSpec.describe SpreeEmerchantpayGenesis::TransactionHelper do
   it 'when generate response with default object' do
     expect(
       described_class.generate_spree_response(ActiveMerchant::Billing::Response.new(true, 'Success'))
-    ).to be_kind_of ActiveMerchant::Billing::Response
+    ).to be_a ActiveMerchant::Billing::Response
   end
 
   it 'when init reference request with invalid arguments' do