+++ spec/models/spree_emerchantpay_genesis/base/data_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe SpreeEmerchantpayGenesis::Base::Data do
   it 'when respond_to_missing' do
-    expect(described_class.new.respond_to?(:missing)).to eq false
+    expect(described_class.new.respond_to?(:missing)).to be false
   end
 end