+++ spec/models/spree_emerchantpay_genesis/data/user_spec.rb
@@ -6,13 +6,13 @@ RSpec.describe SpreeEmerchantpayGenesis::Data::User do
     address = create_address
     user.bill_address_id = address.id
 
-    expect(user.billing_address).to be_kind_of Spree::Address
+    expect(user.billing_address).to be_a Spree::Address
   end
 
   it 'with ship_address_id' do
     address = create_address
     user.ship_address_id = address.id
 
-    expect(user.shipping_address).to be_kind_of Spree::Address
+    expect(user.shipping_address).to be_a Spree::Address
   end
 end